Add String::ellipsize method.
This method is used for trimming a human-readable string to a given
character count.  An ellipsis is used when the string is too long.

This gives a hard limit on the length of the result so that you don't
have to worry about the string being longer with the ellipsis added.
This can be useful if the string is being truncated for insertion into a
database or anywhere where the EXACT length limit needs to be enforced.